+++ src/disasm.cpp
@@ -496,34 +496,43 @@ jl_value_t *jl_dump_function_ir_impl(void *f, char strip_ir_metadata, char dump_
     raw_string_ostream stream(code);
 
     {
+        //RAII will release the struct itself
         std::unique_ptr<jl_llvmf_dump_t> dump(static_cast<jl_llvmf_dump_t*>(f));
-        dump->TSM.withModuleDo([&](Module &m) {
-            Function *llvmf = dump->F;
-            if (!llvmf || (!llvmf->isDeclaration() && !llvmf->getParent()))
-                jl_error("jl_dump_function_ir: Expected Function* in a temporary Module");
+        //RAII will release the module
+        auto TSM = std::unique_ptr<orc::ThreadSafeModule>(unwrap(dump->TSM));
+        //If TSM is not passed in, then the context MUST be locked externally.
+        //RAII will release the lock
+        Optional<orc::ThreadSafeContext::Lock> lock;
+        if (TSM) {
+            lock.emplace(TSM->getContext().getLock());
+        }
+        Function *llvmf = cast<Function>(unwrap(dump->F));
+        if (!llvmf || (!llvmf->isDeclaration() && !llvmf->getParent()))
+            jl_error("jl_dump_function_ir: Expected Function* in a temporary Module");
 
-            LineNumberAnnotatedWriter AAW{"; ", false, debuginfo};
-            if (!llvmf->getParent()) {
-                // print the function declaration as-is
-                llvmf->print(stream, &AAW);
-                delete llvmf;
+        LineNumberAnnotatedWriter AAW{"; ", false, debuginfo};
+        if (!llvmf->getParent()) {
+            // print the function declaration as-is
+            llvmf->print(stream, &AAW);
+            delete llvmf;
+        }
+        else {
+            assert(TSM && TSM->getModuleUnlocked() == llvmf->getParent() && "Passed module was not the same as function parent!");
+            auto m = TSM->getModuleUnlocked();
+            if (strip_ir_metadata) {
+                std::string llvmfn(llvmf->getName());
+                jl_strip_llvm_addrspaces(m);
+                jl_strip_llvm_debug(m, true, &AAW);
+                // rewriting the function type creates a new function, so look it up again
+                llvmf = m->getFunction(llvmfn);
+            }
+            if (dump_module) {
+                m->print(stream, &AAW);
             }
             else {
-                if (strip_ir_metadata) {
-                    std::string llvmfn(llvmf->getName());
-                    jl_strip_llvm_addrspaces(&m);
-                    jl_strip_llvm_debug(&m, true, &AAW);
-                    // rewriting the function type creates a new function, so look it up again
-                    llvmf = m.getFunction(llvmfn);
-                }
-                if (dump_module) {
-                    m.print(stream, &AAW);
-                }
-                else {
-                    llvmf->print(stream, &AAW);
-                }
+                llvmf->print(stream, &AAW);
             }
-        });
+        }
     }
 
     return jl_pchar_to_string(stream.str().data(), stream.str().size());
@@ -1200,10 +1209,11 @@ jl_value_t *jl_dump_function_asm_impl(void *F, char raw_mc, const char* asm_vari
     SmallVector<char, 4096> ObjBufferSV;
     { // scope block
         std::unique_ptr<jl_llvmf_dump_t> dump(static_cast<jl_llvmf_dump_t*>(F));
-        Function *f = dump->F;
+        auto TSM = std::unique_ptr<orc::ThreadSafeModule>(unwrap(dump->TSM));
         llvm::raw_svector_ostream asmfile(ObjBufferSV);
-        assert(!f->isDeclaration());
-        dump->TSM.withModuleDo([&](Module &m) {
+        TSM->withModuleDo([&](Module &m) {
+            Function *f = cast<Function>(unwrap(dump->F));
+            assert(!f->isDeclaration());
             for (auto &f2 : m.functions()) {
                 if (f != &f2 && !f->isDeclaration())
                     f2.deleteBody();
@@ -1217,7 +1227,7 @@ jl_value_t *jl_dump_function_asm_impl(void *F, char raw_mc, const char* asm_vari
             raw_svector_ostream obj_OS(ObjBufferSV);
             if (TM->addPassesToEmitFile(PM, obj_OS, nullptr, CGFT_ObjectFile, false, nullptr))
                 return jl_an_empty_string;
-            dump->TSM.withModuleDo([&](Module &m) { PM.run(m); });
+            TSM->withModuleDo([&](Module &m) { PM.run(m); });
         }
         else {
             MCContext *Context = addPassesToGenerateCode(TM, PM);
@@ -1261,7 +1271,7 @@ jl_value_t *jl_dump_function_asm_impl(void *F, char raw_mc, const char* asm_vari
                 return jl_an_empty_string;
             PM.add(Printer.release());
             PM.add(createFreeMachineFunctionPass());
-            dump->TSM.withModuleDo([&](Module &m){ PM.run(m); });
+            TSM->withModuleDo([&](Module &m){ PM.run(m); });
         }
     }
     return jl_pchar_to_string(ObjBufferSV.data(), ObjBufferSV.size());