+++ stdlib/LibGit2/test/libgit2.jl
@@ -220,6 +220,12 @@ end
     end
 end
 
+@testset "Trace" begin
+    code = "import LibGit2; LibGit2.trace_set(LibGit2.Consts.TRACE_DEBUG); exit(LibGit2.trace_set(0))"
+    p = run(`$(Base.julia_cmd()) --startup-file=no -e $code`, wait=false); wait(p)
+    @test success(p)
+end
+
 # See #21872 and #21636
 LibGit2.version() >= v"0.26.0" && Sys.isunix() && @testset "Default config with symlink" begin
     with_libgit2_temp_home() do tmphome